requirements.txt
streamlit>=1.30.0
google-cloud-firestore>=2.11.0
pandas>=2.0.0
